Use RMM's `DeviceBuffer` for Python allocations
Instead of using RMM's `device_array`, which is planned to go away,
switch to using RMM's `DeviceBuffer`. Since this already supports
`__cuda_array_interface__`, it's straightforward to get a Numba array
that views the allocation. From there only a little work is needed to
cast it to the expected type and shape accordingly.