Fix a critical issue with `parallel_search` functor. (#405)
`parallel_search` was producing the wrong indexes for deciding which coefficients to use for interpolation. This was hidden by the test cases, but exposed via python as I was researching another bug.

This PR resolves the issue with parallel_search so that it produces the correct indexes for interpolation.

I'm improving the testing a bit more before asking for reviews.

Authors:
  - H. Thomson Comer (https://github.com/thomcom)
  - Paul Taylor (https://github.com/trxcllnt)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Paul Taylor (https://github.com/trxcllnt)

URL: https://github.com/rapidsai/cuspatial/pull/405