Refactor haversine_distance to a header-only API (#477)
This is an exploration of what a refactoring of libcuspatial to a header-only API would look like. This PR refactors a single API, `haversine_distance`. The new API is iterator-based, and is completely independent of libcudf and libcudf_test.

The design is based on `std`/Thrust algorithms. It is an iterator-based API, which enables flexibility to use "fancy" iterators such as transform iterators.

There is no support for nulls. We didn't allow `cudf::column` inputs with NULLS in the cuDF-based API anyway.

Authors:
  - Mark Harris (https://github.com/harrism)

Approvers:
  - Michael Wang (https://github.com/isVoid)
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Jake Hemstad (https://github.com/jrhemstad)

URL: https://github.com/rapidsai/cuspatial/pull/477