Fix out of bounds access in spline interpolation (#544)
The cubic spline API supports fitting many splines at once. Each spline's coefficients are stored on construction of the `CubicSpline` object, and when users then specify a set of points at which to evaluate that spline they can specify which spline to evaluate for each of the input points. When a user does not specify that information, the call operator defaults to assuming that only a single spline exists and interpolates all the data using that spline by setting all input points to use spline 0. Currently, however, the code for that creating that default list only making it as long as the number of knots in the spline instead of making it as long as the set of input points to evaluate the spline at. This leads to out of bounds memory accesses when that curve is fit.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cuspatial/pull/544