Add C++ API for `point_linestring_nearest_points` (#658)
This PR adds C++ API of `point_linestring_nearest_points`. The host counterpart of this function is `shapely.ops.nearest_points`. Multiple fields are written when called on this API, including the geometry and part id indicating the geometry that the nearest point is on. In the header only API, the output iterator accepts a zipped iterator made of 4 output iterators. In the column API, a tuple is returned.

The kernel is parallelized per geometry. Each thread computes the nearest point between a pair of multipoint and multilinestring.

Contribute to #646

Authors:
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - H. Thomson Comer (https://github.com/thomcom)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cuspatial/pull/658