`unwrap_pyoptcol` is missing `except` keyword that causes exceptions ignored, fixes function bug (#719)
Fixes a bug in the cdef function when `pyoptcol` is `None` and gets mishandled (raised as error). This error is captured in one of the pytests for `point_linestring_distance`. However, due to the cdef function is missing the `except` keyword, the exception gets ignored by the python interpreter and not reraised.

Authors:
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - H. Thomson Comer (https://github.com/thomcom)

URL: https://github.com/rapidsai/cuspatial/pull/719