Header-only refactoring of trajectory_distances_and_speeds (#769)
Closes #563 

In addition to refactoring to header-only API, this PR changes the implementation to use `thrust_transform` (two input version) instead of `thrust::adjacent_difference`. This simplifies the code a bit and eliminates the need for the `duplicate_first_element_iterator`, which was a bit confusing.

Authors:
  - Mark Harris (https://github.com/harrism)

Approvers:
  - Michael Wang (https://github.com/isVoid)

URL: https://github.com/rapidsai/cuspatial/pull/769