Update pre-commit configuration and CI (#799)
This PR updates the pre-commit configuration and uses pre-commit for all CI style checks. This aligns the configuration closer to that of cudf, but does not add all of the hooks from cudf's list to keep this PR limited in scope.

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - AJ Schmidt (https://github.com/ajschmidt8)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cuspatial/pull/799