Allow None rows in `GeoSeries` and an `align` method to match `GeoPandas` (#760)
This PR allows us to have `GeoSeries` with empty rows containing `None`. This feature is used by GeoPandas frequently with binary operations on `GeoSeries` of different lengths. It also adds an `align` method which works with default arguments for aligning the indices of a `Series` and adding None where rows are missing.

I also updated the internals for `GeoColumn`, `GeoSeries`, and `GeoDataFrame` to handle native `_gather` operations from cudf correctly. This eliminates the need to use `column._data` as a reordering map when `sort_values` is used on a non-geometry column, and should support more hidden cudf operations that we haven't attempted yet.

Closes #762
Closes #761

Authors:
  - H. Thomson Comer (https://github.com/thomcom)

Approvers:
  - Michael Wang (https://github.com/isVoid)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cuspatial/pull/760