Create `pairwise_point_in_polygon` to be used by pairwise `GeoSeries` (#750)
This PR provides two primary outcomes: it adds a `pairwise_point_in_polygon` API in the header-only and cudf-backed C++ functionality, and it modifies the `is_point_in_polygon` kernel to perform boundary exclusion.

The `pairwise_point_in_polygon` API has been added because `GeoPandas.GeoSeries.contains` only supports a pairwise operation and a many-to-one operation, but does not implement the all-pairs approach to point-in-polygon we developed.

Boundary exclusion has been added for two reasons: The main reason is because `GeoPandas.GeoSeries.contains` applies boundary exclusion: points that are in the boundary of a polygon are not considered to be "in" the polygon. The other reason to add boundary exclusion is because our results were previously undefined. Some boundary points were excluded in point-in-polygon, and others were not. This PR resolves that inconsistency.

Authors:
  - H. Thomson Comer (https://github.com/thomcom)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - Michael Wang (https://github.com/isVoid)

URL: https://github.com/rapidsai/cuspatial/pull/750