Fix: Sort Intersection Result by Keys in `intersection_with_duplicate` Test (#857)
This PR sorts the intersection result in `linestring_intersection_with_duplicate` test. If a pair has more than 1 intersection result (e.g. 2 intersecting points) the order of these geometries is non-deterministic. While this doesn't affect the semantic of the result, this makes tests flaky and may break CI. Thus in the tests we have to sort the result before testing. This PR adds the infrastructure to sort the results, as well as rearranging the order to some of the hard coded results.

Authors:
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cuspatial/pull/857