Fix a bug preventing `align` from working with a basic slice. (#850)
This PR fixes a bug with `align` so that slicing a `GeoSeries`, then calling a binpred with slices of the `GeoSeries` with unmatching indexes works. It also adds a few more valid tests.

As of this PR one test still fails, I'm working on it but it feels like a rabbit hole. I'm inclined to leave the case for later:
The case that doesn't work is the hardest case wherein two `GeoSeries` have different length as well as indexes in one that is not present in the other.

Authors:
  - H. Thomson Comer (https://github.com/thomcom)

Approvers:
  - Michael Wang (https://github.com/isVoid)

URL: https://github.com/rapidsai/cuspatial/pull/850