: Fix linestring link in readme (#1003)
This fixes the link text for the linestring intersection function in the readme. In my haste to get the readme in, I failed to notice the link ate the link-text instead of auto updating like it usually does in the GitHub web ui.

Authors:
  - Ben Jarmak (https://github.com/jarmak-nv)

Approvers:
  - Michael Wang (https://github.com/isVoid)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cuspatial/pull/1003