Reduced equals time and fixed a bug. (#1051)
Closes https://github.com/rapidsai/cuspatial/issues/1014

This PR simply reduces the size of some of the equals tests from 10000 to 100. Long run time had to do with serializing 10000x Shapely objects.

It also fixes a bug in `MultiPointMultiPointEquals` that I guess was able to exist due to the size of the old test and the seed of the random generator.

Authors:
  - H. Thomson Comer (https://github.com/thomcom)

Approvers:
  - Michael Wang (https://github.com/isVoid)

URL: https://github.com/rapidsai/cuspatial/pull/1051