Bug fix in `pairwise_linestring_intersection` (#1069)
This PR fixes a bug in `pairwise_linestring_intersection`, where `point_flags` is incorrectly used to remove points in points array due to leftover results from find_duplicate_points.

Previously, the code incorrectly  assumed that `point_flags` will always be written to prior to `merge_point_on_segment` kernel, however, this may not be true if the result segment array is empty, and the kernel launch could be evaded, leading to that the second `remove_if` operates on the "leftover" from the previous `point_flag` ramnant.

This fix is to only run segment cleanup kernels if there is at least one segment result in the output. This not only fixes #1067 , but also serve as an optimization to avoid unnecessary kernel launching.

It should be pointed out that the result from python isn't exactly 1-1 matching with geopandas, as geopandas returns
```python
0    MULTIPOINT (1.00000 1.00000, 0.00000 0.00000)
1    MULTIPOINT (1.00000 1.00000, 0.00000 0.00000)
2    MULTIPOINT (1.00000 1.00000, 0.00000 0.00000)
```
While cuspatial returns the result in 6 point rows with geometry offsets.

Authors:
  - Michael Wang (https://github.com/isVoid)
  - Mark Harris (https://github.com/harrism)

Approvers:
  - H. Thomson Comer (https://github.com/thomcom)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cuspatial/pull/1069