Fix a bug in point-in-polygon kernel: if the point is collinear with an edge, result is asserted false (#1108)
Fixes #1103, current algorithm tests if the test point is collinear with an edge of the polygon, the point is asserted to be on the edge. This is not true, because for a point to be on the edge, the point also needs to be within the range where the edge covers. Collinearity test only test if the point is covered by the line that the edge coincides.

This PR fixes this bug by adding additional tests to guarantee that the point is on an edge iff the point is collinear with the line where the edge coincides as well as the x coordinate of the point is within the closed range of the edge's x coordinates.

This PR also fixes an additional bug where the col-linearity flag is not reset after each iteration of a ring.

Authors:
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/cuspatial/pull/1108