WGS84 <--> UTM projection header-only API (#1191)
Implements WGS to UTM projection (either direction) in CUDA C++. This is based on the code from PROJ, and that library is used for comparison in the tests.

To reviewers, I would start by looking over the test: wgs_to_utm_test.cu. This uses `projection_factories.hpp`, so review that next, then look at `projection.cuh` and `projection_parameters.cuh`. The projection pipeline is applied using a `thrust::transform()` where the operator is a `pipeline` made up of a sequence of `operation`s (operation.cuh).  A pipeline can be applied forward or inverse. The operations for UTM projection are all in include/cuproj/operation. 

Note that the dispatch of operations is a little clunky (switch statement in a device function). This is due to limitations of creating arrays of objects on the host for virtual dispatch on the device. This bears more experimentation in the future to come up with a better approach, but this works fine for now.

~TODO:~

- [x] Design a way to build pipelines of device operations dynamically and run on GPU
- [x] Documentation
- [x] More testing

Benchmark and Python API are left to follow-up PRs.

Authors:
  - Mark Harris (https://github.com/harrism)

Approvers:
  - Michael Wang (https://github.com/isVoid)
  - H. Thomson Comer (https://github.com/thomcom)

URL: https://github.com/rapidsai/cuspatial/pull/1191