Enable pytest failures on warnings from cudf (Ensure values being set are casted to the object's type) (#1358)
In 24.04, cudf issues a `FutureWarning` when setting a value to a column would change the data type of the original column. The fix involves casting the original column to the value's type first or vice versa. Could use a second eye on the correct direction to cast if a data type was used intentionally.

Additionally the test suite will now fail when a warning comes from cudf as apart of https://github.com/rapidsai/build-planning/issues/26

Authors:
  - Matthew Roeschke (https://github.com/mroeschke)

Approvers:
  - Mark Harris (https://github.com/harrism)
  - H. Thomson Comer (https://github.com/thomcom)

URL: https://github.com/rapidsai/cuspatial/pull/1358