Fix libcuproj doc example and make tests consistent. (#1503)
This fixes libcuproj documentation example inconsistencies raised in #1500.

It removes the (outdated) example from the doxygen `main_page` and replaces it with a link to the readme.

It also makes the type definitions simpler in the README example and makes the Test code match the example.

Authors:
  - Mark Harris (https://github.com/harrism)
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - Michael Wang (https://github.com/isVoid)

URL: https://github.com/rapidsai/cuspatial/pull/1503